# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Weather API",
      "version": "1.0.0",
      "title": "Weather API 3"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "tags": [
      {
        "name": "weather",
        "description": "Weather data management"
      }
    ],
    "paths": {
      "/api/weather": {
        "get": {
          "tags": ["weather"],
          "summary": "Get weather data",
          "parameters": [
            {
              "name": "station_id",
              "in": "query",
              "description": "Weather station ID",
              "required": false,
              "type": "string"
            },
            {
              "name": "start_date",
              "in": "query",
              "description": "Start date",
              "required": false,
              "type": "string",
              "format": "date"
            },
            {
              "name": "end_date",
              "in": "query",
              "description": "End date",
              "required": false,
              "type": "string",
              "format": "date"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "station_id": {
                      "type": "string"
                    },
                    "date": {
                      "type": "string",
                      "format": "date"
                    },
                    "max_temp": {
                      "type": "integer"
                    },
                    "min_temp": {
                      "type": "integer"
                    },
                    "precipitation": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "/api/weather/stats": {
        "get": {
          "tags": ["weather"],
          "summary": "Get weather statistics",
          "parameters": [
            {
              "name": "station_id",
              "in": "query",
              "description": "Weather station ID",
              "required": false,
              "type": "string"
            },
            {
              "name": "year",
              "in": "query",
              "description": "Year",
              "required": false,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response",
              "schema": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "station_id": {
                      "type": "string"
                    },
                    "year": {
                      "type": "integer"
                    },
                    "avg_max_temp": {
                      "type": "number"
                    },
                    "avg_min_temp": {
                      "type": "number"
                    },
                    "total_precipitation": {
                      "type": "number"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
  